requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
chardet==4.0.0
geographiclib==1.50
geopy==2.1.0
idna==2.10
llvmlite==0.36.0
numba==0.53.1
numpy==1.20.2
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
slack-sdk==3.4.2
timezonefinder==5.2.0
tzlocal==2.1
urllib3==1.26.4
